Convert to plugin structure
