+++ .eslintrc
@@ -57,6 +57,7 @@
     "no-plusplus": 0,
     "no-restricted-syntax": 0,
     "no-underscore-dangle": 0,
-    "no-unused-vars": ["error", { "varsIgnorePattern": "^_", "argsIgnorePattern": "^_", "caughtErrors": "none" }]
+    "no-unused-vars": ["error", { "varsIgnorePattern": "^_", "argsIgnorePattern": "^_", "caughtErrors": "none" }],
+    "consistent-return": ["warn", { "treatUndefinedAsUnspecified": false }]
   }
 }
