+++ .travis.yml
@@ -1,8 +1,8 @@
 language: node_js
 node_js:
-  - "8"
   - "10"
   - "12"
+  - "14"
 
 install:
   - npm ci
