+++ package.json
@@ -13,34 +13,34 @@
     "json"
   ],
   "engines": {
-    "node": ">= 10.15.3"
+    "node": ">= 12"
   },
   "dependencies": {
     "debug": "^4.1.1",
     "ramda": "^0.27.0",
-    "uttori-utilities": "^2.1.0"
+    "uttori-utilities": "^2.1.1"
   },
   "devDependencies": {
-    "ava": "^3.2.0",
-    "coveralls": "^3.0.9",
+    "ava": "^3.8.2",
+    "coveralls": "^3.1.0",
     "docsify-cli": "^4.4.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-plugin-ava": "^10.0.1",
-    "eslint-plugin-import": "^2.20.1",
+    "eslint": "^7.1.0",
+    "eslint-config-airbnb": "^18.1.0",
+    "eslint-plugin-ava": "^10.3.0",
+    "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-no-inferred-method-name": "^2.0.0",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-optimize-regex": "^1.1.7",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-optimize-regex": "^1.2.0",
     "eslint-plugin-ramda": "^2.5.1",
-    "eslint-plugin-react": "^7.18.3",
-    "eslint-plugin-react-hooks": "^2.3.0",
+    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react-hooks": "^4.0.4",
     "eslint-plugin-security": "^1.4.0",
-    "eslint-plugin-unicorn": "^16.0.0",
-    "eslint-plugin-xss": "^0.1.9",
+    "eslint-plugin-unicorn": "^20.1.0",
+    "eslint-plugin-xss": "^0.1.10",
     "jsdoc-to-markdown": "^5.0.3",
-    "nyc": "^15.0.0",
-    "sinon": "^8.1.1",
+    "nyc": "^15.0.1",
+    "sinon": "^9.0.2",
     "supertest": "^4.0.2",
     "uttori-document": "^1.4.1"
   },
@@ -70,7 +70,7 @@
     "exclude": []
   },
   "scripts": {
-    "docs": "jsdoc2md --private --example-lang js --template README.hbs src/index.js > README.md; cp README.md docs/README.md",
+    "docs": "jsdoc2md --private --example-lang js --template README.hbs src/storage-provider.js > README.md; cp README.md docs/README.md",
     "docs-serve": "docsify init ./docs && docsify serve ./docs",
     "lint": "eslint .",
     "report": "nyc report --reporter=html",
